J-87: + shownote input list component
