+++ frontend/src/components/configurate-episode/components/create-episode-form/styles.module.scss
@@ -10,37 +10,3 @@
   gap: 15px;
   border: none;
 }
-
-.shownotesTitle {
-  display: flex;
-  align-items: center;
-  font-size: 12px;
-  font-weight: 500;
-
-  & button {
-    margin-left: 16px;
-    padding: 2px 10px;
-    font-weight: 600;
-  }
-}
-
-.shownotesList {
-  margin: 0;
-  list-style: none;
-
-  & button {
-    padding: 4px 10px;
-    font-weight: 600;
-  }
-
-  &Item {
-    margin-bottom: 25px;
-  }
-}
-
-.errorWrapper {
-  font-size: 12px;
-  font-weight: 500;
-  color: var(--color-button-background);
-  margin-bottom: 7px;
-}
